core: add byte HWM transport lanes

+++ core/src/api/monitoring/monitor_socket_api.cpp
@@ -83,7 +83,7 @@ void *open_socket_monitor_with_handler_internal (void *s_,
         return NULL;
     }
     monitor_socket_base->set_auto_hwm_policy_enabled (false);
-    const int monitor_hwm = 4096;
+    const uint64_t monitor_hwm = 4096;
     (void) monitor_socket_base->setsockopt (ZLINK_INTERNAL_OPT_SNDHWM, &monitor_hwm,
                                             sizeof (monitor_hwm));
     (void) monitor_socket_base->setsockopt (ZLINK_INTERNAL_OPT_RCVHWM, &monitor_hwm,
