+++ core/src/runtime/core/ctx.hpp
@@ -116,7 +116,7 @@ class ctx_t ZLINK_FINAL
     int auto_hwm_recalculate_now ();
     zlink_auto_hwm_profile_t auto_hwm_profile () const;
     bool auto_hwm_enabled () const;
-    int auto_hwm_msg_unit_bytes () const;
+    uint64_t auto_hwm_msg_unit_bytes () const;
 
   private:
     friend class ctx_termination_test_access_t;
